% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{abs.df}
\alias{abs.df}
\title{A data frame with candidates' HLA antibodies}
\usage{
abs.df(candidates = candidates.df(n = 10), n.seed = 3)
}
\arguments{
\item{n.seed}{a numeric seed that will be used for random number generation.}
}
\value{
A data frame with \{ID} and HLA antibodies \code{Abs}.
}
\description{
Returns a data frame with transplant candidates' HLA antibodies obtained according to theirs cPRA values and HLA typing.
}
\examples{
abs.df(candidates = candidates.df(n=10), n.seed = 3)
}
