% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agA}
\alias{agA}
\title{A vector with HLA-A.}
\format{
A character vector with length 20.
}
\usage{
agA
}
\description{
A character vector with antigens for HLA-A locus obtained from \code{hla} dataset.
}
\concept{data}
\keyword{datasets}
