% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agA_MNDP}
\alias{agA_MNDP}
\title{A vector with HLA-A from MNDP.}
\format{
A character vector with length 21.
}
\usage{
agA_MNDP
}
\description{
A character vector with  HLA-A locus antigens, obtained from \code{MNDPhaps} dataset.
}
\keyword{datasets}
