% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agB}
\alias{agB}
\title{A vector with HLA-B.}
\format{
A character vector with length 34.
}
\usage{
agB
}
\description{
A character vector with antigens for HLA-B locus obtained from \code{hla} dataset.
}
\concept{data}
\keyword{datasets}
