% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agB_MNDP}
\alias{agB_MNDP}
\title{A vector with HLA-B from MNDP.}
\format{
A character vector with length 41.
}
\usage{
agB_MNDP
}
\description{
A character vector with  HLA-B locus antigens, obtained from \code{MNDPhaps} dataset.
}
\keyword{datasets}
