% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agDR}
\alias{agDR}
\title{A vector with HLA-DR.}
\format{
A character vector with length 13.
}
\usage{
agDR
}
\description{
A character vector with antigens for HLA-DR locus obtained from \code{hla} dataset.
}
\keyword{datasets}
