% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agDR_MNDP}
\alias{agDR_MNDP}
\title{A vector with HLA-DR from MNDP.}
\format{
A character vector with length 13.
}
\usage{
agDR_MNDP
}
\description{
A character vector with  HLA-DR locus antigens, obtained from \code{MNDPhaps} dataset.
}
\concept{data}
\keyword{datasets}
