% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{aGFR}
\alias{aGFR}
\title{Gives a eGFR by age}
\usage{
aGFR(age = 43, seed.number = NA)
}
\arguments{
\item{age}{An integer for age (values between 1 and 99).}

\item{seed.number}{Seed for pseudo random number generator.}
}
\value{
A value from a normal distribution.
}
\description{
Returns a value for the Estimated Glomerular Filtration Rate
(eGFR) by age as described by https://www.kidney.org/atoz/content/gfr.
}
\examples{
aGFR(age = 43, seed.number = NA)
}
\concept{clinical_parameters}
