% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo.R
\name{antbs}
\alias{antbs}
\title{samples HLA antibodies}
\usage{
antbs(
  cA = c("2", "29"),
  cB = c("7", "15"),
  cDR = c("4", "7"),
  cPRA = 85,
  origin = "PT",
  n_seed = 3
)
}
\arguments{
\item{cA}{candidate's HLA-A typing}

\item{cB}{candidate's HLA-B typing}

\item{cDR}{candidate's HLA-DR typing}

\item{cPRA}{candidate's cPRA value}

\item{origin}{A character value from options: 'API', 'AFA', 'CAU' and 'HIS'}

\item{n_seed}{a numeric seed that will be used for random number generation.}
}
\value{
a character vector with HLA abs.
}
\description{
creates a sample of HLA antibodies (abs) for a given candidate
according with a cPRA value.
}
\examples{
antbs(cA = c('2','29'), cB = c('7','15'), cDR = c('4','7'),
cPRA = 85, origin = 'PT', n_seed = 3)
}
\concept{histocompatibility}
