% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{candidates.df}
\alias{candidates.df}
\title{A data frame with donors characteristics}
\usage{
candidates.df(
  n = 1000,
  replace = TRUE,
  probs.abo = c(0.43, 0.03, 0.08, 0.46),
  probs.cpra = c(0.7, 0.1, 0.1, 0.1),
  lower = 18,
  upper = 75,
  mean = 45,
  sd = 15,
  prob.dm = 0.12,
  uk = FALSE,
  n.seed = 3
)
}
\arguments{
\item{n}{An integer to define the number of rows}

\item{replace}{A logical value for sampling with replacement}

\item{probs.abo}{A vector with the probabilities for blood group A, AB, B and O (in this order). The sum of the probabilities must be equal to one.}

\item{probs.cpra}{A vector with the probabilities for cPRA groups 0\%, 1\%-50\%, 51\%-84\%, 85\%-100\% (in this order). The sum of the probabilities must be equal to one.}

\item{lower}{An integer for ages' lower limit.}

\item{upper}{An integer for ages' upper limit.}

\item{mean}{A value for mean age's distribution.}

\item{sd}{A value for standard deviation age's distribution.}

\item{prob.dm}{A value for the probability of having Diabetes Mellitus}

\item{uk}{A logical value, if TRUE is also computed the Donor Risk Index (DRI)}

\item{n.seed}{a numeric seed that will be used for random number generation.}
}
\value{
A data frame with HLA typing, blood group and truncated ages for a simulated group of transplant donors.
}
\description{
Returns a data frame for simulated donors clinical and demographic characteristics.
}
\examples{
candidates.df(n = 1000, replace = TRUE, probs = c(0.4658, 0.0343, 0.077, 0.4229), lower=18, upper=75, mean = 55, sd = 15, prob.dm = 0.12, uk = FALSE, n.seed = 3)
}
