% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{candidates_df}
\alias{candidates_df}
\title{A data frame with donors characteristics}
\usage{
candidates_df(
  n = 10,
  replace = TRUE,
  origin = "PT",
  probs_abo = c(0.43, 0.03, 0.08, 0.46),
  probs_cpra = c(0.7, 0.1, 0.1, 0.1),
  lower = 18,
  upper = 75,
  mean = 45,
  sd = 15,
  prob_dm = 0.12,
  uk = FALSE,
  n_seed = 3
)
}
\arguments{
\item{n}{An integer to define the number of rows}

\item{replace}{A logical value for sampling with replacement}

\item{origin}{A character value for population origin from options: 'PT', 'API', 'AFA', 'CAU' and 'HIS'}

\item{probs_abo}{A vector with the probabilities for blood group A, AB, B and O (in this order). The sum of the probabilities must be equal to one.}

\item{probs_cpra}{A vector with the probabilities for cPRA groups 0\%, 1\%-50\%, 51\%-84\%, 85\%-100\% (in this order). The sum of the probabilities must be equal to one.}

\item{lower}{An integer for ages' lower limit.}

\item{upper}{An integer for ages' upper limit.}

\item{mean}{A value for mean age's distribution.}

\item{sd}{A value for standard deviation age's distribution.}

\item{prob_dm}{A value for the probability of having Diabetes Mellitus}

\item{uk}{A logical value, if TRUE is also computed the Donor Risk Index (DRI)}

\item{n_seed}{a numeric seed that will be used for random number generation.}
}
\value{
A data frame with HLA typing, blood group, truncated ages, time on dialysis (in months), cPRA, Tier, MS and RRI (those last 3, only when uk = TRUE) for a simulated group of transplant candidates.
}
\description{
Returns a data frame for simulated donors clinical and
demographic characteristics.
}
\examples{
candidates_df(n = 10, replace = TRUE, origin = 'PT',
probs_abo = c(0.43, 0.03, 0.08, 0.46), probs_cpra = c(0.7, 0.1, 0.1, 0.1),
lower=18, upper=75, mean = 45, sd = 15,
prob_dm = 0.12,
uk = FALSE, n_seed = 3)
}
\concept{generate_data}
