% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{D10K}
\alias{D10K}
\title{A pool of 10.0000 donors.}
\format{
A data frame with 37993 rows and 9 variables:
\describe{
\item{ID}{Identifier}
\item{bg}{ABO blood group}
\item{A1}{HLA-A allele 1}
\item{A2}{HLA-A allele 2}
\item{B1}{HLA-B allele 1}
\item{B2}{HLA-B allele 2}
\item{DR1}{HLA-DRB1 allele 1}
\item{DR2}{HLA-DRB1 allele 2}
\item{age}{age in years}
...
}
}
\source{
\url{https://www.slideshare.net/balima78/lima-2013}
}
\usage{
D10K
}
\description{
A synthetic dataset with information for 10.000 donors, generated with \code{donors.df()} and \code{set.seed(3)} .
}
\keyword{datasets}
