% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{donors_df}
\alias{donors_df}
\title{A data frame with donors caracheristics}
\usage{
donors_df(
  n = 10,
  replace = TRUE,
  origin = "PT",
  probs = c(0.4658, 0.0343, 0.077, 0.4229),
  lower = 18,
  upper = 75,
  mean.age = 60,
  sd.age = 12,
  uk = FALSE,
  seed.number = 3
)
}
\arguments{
\item{n}{An integer to define the number of rows}

\item{replace}{A logical value for sampling with replacement}

\item{origin}{A character value for population origin from options: 'PT', 'API', 'AFA', 'CAU' and 'HIS'}

\item{probs}{A vector with the probabilities for blood group A, AB, B and O (in this order). The sum of the probabilities must be equal to one.}

\item{lower}{An integer for ages' lower limit.}

\item{upper}{An integer for ages' upper limit.}

\item{mean.age}{A value for mean on age's distribution.}

\item{sd.age}{A value for standard deviation on age's distribution.}

\item{uk}{A logical value, if TRUE is also computed the Donor Risk Index (DRI)}

\item{seed.number}{a numeric seed that will be used for random number generation.}
}
\value{
A data frame with HLA typing, blood group and truncated ages for a simulated group of transplant donors.
}
\description{
Returns a data frame for simulated donors clinical and
demographic caracheristics.
}
\examples{
donors_df(n = 10, replace = TRUE, origin = 'PT',
probs = c(0.4658, 0.0343, 0.077, 0.4229),
lower=18, upper=75, mean.age = 55, sd.age = 15,
uk = FALSE, seed.number = 3)
}
\concept{generate_data}
