% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hla}
\alias{hla}
\title{HLA typing.}
\format{
A data frame with 37993 rows and 9 variables:
\describe{
\item{ID}{Identifier}
\item{A1}{HLA-A allele 1}
\item{A2}{HLA-A allele 2}
\item{B1}{HLA-B allele 1}
\item{B2}{HLA-B allele 2}
\item{C1}{HLA-C allele 1}
\item{C2}{HLA-C allele 2}
\item{DRB11}{HLA-DRB1 allele 1}
\item{DRB12}{HLA-DRB1 allele 2}
...
}
}
\source{
\url{https://www.slideshare.net/balima78/lima-2013}
}
\usage{
hla
}
\description{
A dataset containing HLA typing from 37993 unrelated voluntary bone marrow donors from Nothern Portugal, recruited between 2010 and 2011.
}
\keyword{datasets}
