% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{hla_sample}
\alias{hla_sample}
\title{A data frame with HLA typing}
\usage{
hla_sample(n, replace)
}
\arguments{
\item{n}{An integer to define the number of rows}

\item{replace}{A logical value for sampling with replacement}
}
\value{
A data frame
}
\description{
Returns a data frame with HLA-A, -B, -C and -DRB1 typing
}
\examples{
hla_sample(n = 1000, replace = TRUE)
}
