% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{hla_sample_mndp}
\alias{hla_sample_mndp}
\title{A data frame with HLA haplotypes}
\usage{
hla_sample_mndp(n, replace, origin)
}
\arguments{
\item{n}{An integer to define the number of rows}

\item{replace}{A logical value for sampling with replacement}

\item{origin}{A character value from options: 'API', 'AFA', 'CAU' and 'HIS'}
}
\value{
A data frame
}
\description{
Returns a data frame with HLA-A, -B and -DRB1 haplotypes from MNDP by race and sampled by their frequencies
}
\examples{
hla_sample_mndp(n = 1000, replace = TRUE, origin = 'CAU')
}
\concept{clinical_parameters}
