% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{matchability}
\alias{matchability}
\title{Matchability from D10K}
\usage{
matchability(
  cABO = "A",
  cPRA = 85,
  cA = c("2", "29"),
  cB = c("7", "15"),
  cDR = c("4", "7"),
  n.seed = 3
)
}
\arguments{
\item{cABO}{A character from 'A', 'B', 'AB', 'O'}

\item{cPRA}{candidate's cPRA value}

\item{cA}{candidate's HLA-A typing}

\item{cB}{candidate's HLA-B typing}

\item{cDR}{candidate's HLA-DR typing}

\item{n.seed}{a numeric seed that will be used for random number generation.}
}
\value{
mmA number of HLA-A mismatchs
}
\description{
Computes the number donors on dataset D10K that are a match to a given transplant candidate. A sample of D10K is selected according to cPRA value, and donors ABO identical and HLA mismatch level 1 or 2 (0 DR or (1 DR and 0 B)) are filtered.
}
\examples{
matchability(dA = c('1','2'), dB = c('5','7'), dDR = c('1','4'), cA = c('1','2'), cB = c('03','15'), cDR = c('04','07'))
}
