% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo.R
\name{mmHLA}
\alias{mmHLA}
\title{number of HLA mismatchs}
\usage{
mmHLA(
  dA = c("1", "2"),
  dB = c("5", "7"),
  dDR = c("1", "4"),
  cA = c("1", "2"),
  cB = c("3", "15"),
  cDR = c("4", "7")
)
}
\arguments{
\item{dA}{donor's HLA-A typing}

\item{dB}{donor's HLA-B typing}

\item{dDR}{donor's HLA-DR typing}

\item{cA}{candidate's HLA-A typing}

\item{cB}{candidate's HLA-B typing}

\item{cDR}{candidate's HLA-DR typing}
}
\value{
mmA number of HLA-A mismatchs between \code{dA} and \code{cA};
mmB number of HLA-B mismatchs between \code{dB} and \code{cB};
mmDR number of HLA-DR mismatchs between \code{dDR} and \code{cDR};
and mmHLA as the sum of mmA + mmB + mmDR
}
\description{
Computes the number of HLA mismatchs between one donor and one candidate
}
\examples{
mmHLA(dA = c('1','2'), dB = c('5','7'), dDR = c('1','4'),
cA = c('1','2'), cB = c('03','15'), cDR = c('04','07'))
}
\concept{histocompatibility}
