% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MNDPhaps}
\alias{MNDPhaps}
\title{NMDP HLA frequencies.}
\format{
A data frame with 8696 rows and 7 variables:
\describe{
\item{A}{HLA-A allele}
\item{B}{HLA-B allele}
\item{DR}{HLA-DRB1 allele}
\item{api}{haplotye frequencies for Asian / Pacific Islander}
\item{afa}{haplotye frequencies for African American / Black}
\item{cau}{haplotye frequencies for Withe Caucasian}
\item{his}{haplotye frequencies for Hispanic}
...
}
}
\source{
\url{https://bioinformatics.bethematchclinical.org/hla-resources/haplotype-frequencies/a-b-drb1-224-haplotype-frequencies/}
}
\usage{
MNDPhaps
}
\description{
A dataset containing Frequencies of HLA-A, -B, -DR haplotypes estimated (Expectation Maximization) by race/ethnicity from a subset of donors in the NMDP registry.
}
\keyword{datasets}
