% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simK.R
\docType{package}
\name{simK}
\alias{simK}
\title{simK: A package for data simulation of donors and wait list candidates for kidney transplantation}
\description{
The simK package provides functions for data simulation of clinical and demographic features from cadaveric donors and wait list transplant candidates.
}
\section{simK functions}{

The \code{simK} main functions: \code{candidates.df()}, \code{donors.df()}, \code{abs.df()}
}

