% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo.R
\name{vpra}
\alias{vpra}
\title{virtual PRA}
\usage{
vpra(abs = c("A1", "A2", "B5", "DR4"), donors = D10K)
}
\arguments{
\item{abs}{A character vector with HLA antibodies.}

\item{donors}{A dataframe with HLA typing for a pool of donors.}
}
\value{
a percentual value for vPRA
}
\description{
Computes virtual PRA (vPRA) form HLA-A, -B, -DR loci.
}
\examples{
vpra(abs = c('A1','A2','B5','DR4'), donors = D10K)
}
\concept{histocompatibility}
