% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{abs.df}
\alias{abs.df}
\title{A data frame with candidates' HLA antibodies}
\usage{
\method{abs}{df}(candidates = candidates.df(n = 10), n.seed = 3)
}
\arguments{
\item{candidates}{A dataframe with \code{ID}, HLA typing (column names: \code{A1}, \code{A2}, \code{B1}, \code{B2}, \code{DR1}, \code{DR2}) and cPRA value (column name: \code{cPRA}), for a group of transplant candidates.}

\item{n.seed}{a numeric seed that will be used for random number generation.}
}
\value{
A data frame with \code{ID} and HLA antibodies \code{Abs}.
}
\description{
Returns a data frame with transplant candidates' HLA antibodies obtained according to theirs cPRA values and HLA typing.
}
\examples{
abs.df(candidates = candidates.df(n=10), n.seed = 3)
}
