% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{ages}
\alias{ages}
\title{Builds a vector with truncated ages}
\usage{
ages(n = 100, lower = 18, upper = 75, mean = 55, sd = 15, seed.number = 123)
}
\arguments{
\item{n}{An integer to define the length of the returned vector.}

\item{lower}{An integer for ages' lower limit.}

\item{upper}{An integer for ages' upper limit.}

\item{mean}{A value for mean's distribution.}

\item{sd}{A value for standar deviation's distribution.}

\item{seed.number}{Seed for pseudo random number generator.}
}
\value{
A vector length \code{n} with normal distributed truncated ages.
}
\description{
Returns a vector with ages with a normal distribution truncated
by upper and upper limits.
}
\examples{
ages(n = 100, lower=18, upper=75, mean = 55, sd = 15, seed.number = 123)
}
\concept{demographic_parameters}
