% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\name{hla_sample}
\alias{hla_sample}
\title{A data frame with HLA typing}
\usage{
hla_sample(n, replace, origin)
}
\arguments{
\item{n}{An integer to define the number of rows}

\item{replace}{A logical value for sampling with replacement}

\item{origin}{A character value from options: 'PT', 'API', 'AFA', 'CAU' and 'HIS'}
}
\value{
A data frame
}
\description{
Returns a data frame with HLA-A, -B and -DRB1 typing
}
\examples{
hla_sample(n = 1000, replace = TRUE, origin = 'PT')
}
\concept{clinical_parameters}
